public int[] getSerializedFullObservationZZ(int ZLevelScene, int ZLevelEnemies) method added in to the MarioEnvironment class. Unit test added too.
Returns a solid array consisting of: levelScene, enemies and marioState.
The size of the returned array is: receptiveFieldWidth*receptiveFieldHeight*2+11.

+++ src/ch/idsia/benchmark/mario/environments/Environment.java
@@ -96,7 +96,7 @@ public byte[][] getEnemiesObservationZ(int ZLevelEnemies);
 
 // OBSERVATION FOR AmiCo Agents
 
-public float[] getSerializedFullObservationZZ(int ZLevelScene, int ZLevelEnemies);
+public int[] getSerializedFullObservationZZ(int ZLevelScene, int ZLevelEnemies);
 
 /**
  * Serializes the LevelScene observation from 22x22 byte array to a 1x484 byte array
